Cleaning up App.js
